[Abilities] Made the rechargespeed depending on attributes
This allows a lot more flexibility and makes stuff like magical items a
lot easier to implement. We will also use this for the attack system in
the future. So a attack (abilitiy) would simply depend on some value
like Agility (or a derived attribute from it). Which sets the recharge
speed.

The recharge speed is the modified value of the attribute per game tick.