Do not remove persistet items at startup
If persisting of items is disabled they are not saved in
the database anyway.

So the only case where this happened was if the config was
changed and the server restarted. In which case it should
not matter to recreate them and let the decay normally.

This helps my target to remove dependency on configuration
for the storage module.