Removed the resource handler
The physfs layer was not really required. It mapped together the current
working directory, the world data and the server install dir.

The mapping of the working directory only creates hardly memorizable
issues when running from the wrong directory, the world data path can
simply be accessed from the settings and the server install dir should
only ever be used to load the scripting library.

However even with the scripting library always getting loaded from the
install dir you can still overwrite it in scripts if you which. So no
functionallity is lost (only a bunch of code).